+++ packages/fighting-design/list/src/interface.d.ts
@@ -1,6 +1,1 @@
 export type { ListProps } from './props'
-
-/**
- * 不同尺寸类型
- */
-export type ListSize = 'large' | 'middle' | 'small'