+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- 修复 `f-list` 和 `f-list-item` 组件 bug
+- 新增 `f-list` 组件 `zebra-color` 配置项
+- 新增 `f-list` 组件 `size` 配置项新增 `mini` 可选项
+
 ## 0.9.0 (2022-12-07)
 
 - 完善 `f-menu` 组件